requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
pillow